+++ extensions/menubar-weather/src/utils/weather-utils.ts
@@ -4,7 +4,8 @@ import { isEmpty } from "./common-utils";
 import { getOpenMeteoLocation, getOpenMeteoWeather } from "./axios-utils";
 import { GeoLocation } from "../types/types";
 
-export const { cityName, tempUnits, precipitationUnits, longitude, latitude } = getPreferenceValues<Preferences>();
+export const { cityName, tempUnits, precipitationUnits, longitude, latitude, showSun, showLocation, showForecast } =
+  getPreferenceValues<Preferences>();
 
 export async function getGeoLocation() {
   const geoLocation = await getOpenMeteoLocation();