+++ extensions/menubar-weather/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Fix bug] - {PR_MERGE_DATE}
 
 - Fix the bug of crashing without network
+- Add Preferences **Additional Info**: Whether to show sunrise and sunset, location and 7-day weather forecast
 
 ## [Add forecast] - 2022-11-24